pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdoc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "streamlit-authenticator>=0.4.2",
    "watchdog>=6.0.0",
]

[tool.setuptools.packages.find]
include = ["markdoc*"]
exclude = ["pages*"]
